@@ -1,45 +1,49 @@
 {
-  "name": "glitch-in-bio",
-  "version": "0.0.3",
-  "description": "Your very own link-in-bio app.",
-  "main": "index.html",
-  "scripts": {
-    "start": "vite",
-    "build": "vite build",
-    "serve": "vite preview",
-    "release": "auto shipit"
-  },
-  "devDependencies": {
-    "@auto-it/all-contributors": "^10.32.2",
-    "@auto-it/first-time-contributor": "^10.32.2",
-    "@auto-it/git-tag": "^10.32.3",
-    "@auto-it/released": "^10.32.2",
-    "@glitchdotcom/vite-plugin-handlebars": "^1.6.2",
-    "auto": "^10.32.2",
-    "vite": "^2.5.0",
-    "nunjucks": "^3.2.3"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead"
-  ],
-  "engines": {
-    "node": "14.x"
-  },
-  "repository": {
-    "url": "https://glitch.com/edit/#!/glitch-in-bio"
-  },
-  "glitch": {
-    "projectType": "generated_static"
-  },
-  "author": "Keith Kurson <keith@glitch.com>",
-  "auto": {
-    "plugins": [
-      "all-contributors",
-      "first-time-contributor",
-      "released",
-      "git-tag"
-    ],
-    "onlyPublishWithReleaseLabel": true
-  }
+	"name": "glitch-in-bio",
+	"version": "0.0.3",
+	"description": "Your very own link-in-bio app.",
+	"main": "index.html",
+	"scripts": {
+		"start": "vite",
+		"build": "vite build",
+		"serve": "vite build && vite preview --port 8085"
+	},
+	"devDependencies": {
+		"@auto-it/all-contributors": "^10.32.2",
+		"@auto-it/first-time-contributor": "^10.32.2",
+		"@auto-it/git-tag": "^10.32.3",
+		"@auto-it/released": "^10.32.2",
+		"@glitchdotcom/vite-plugin-handlebars": "^1.6.2",
+		"@vitejs/plugin-basic-ssl": "^2.1.0",
+		"auto": "^10.32.2",
+		"i": "^0.3.7",
+		"npm": "^11.4.2",
+		"nunjucks": "^3.2.3",
+		"vite": "^7.0.2",
+		"vite-plugin-handlebars": "^2.0.0"
+	},
+	"browserslist": [
+		">0.2%",
+		"not dead"
+	],
+	"type": "module",
+	"engines": {
+		"node": "21.x"
+	},
+	"repository": {
+		"url": "https://glitch.com/edit/#!/glitch-in-bio"
+	},
+	"glitch": {
+		"projectType": "generated_static"
+	},
+	"author": "Aram Zucker-Scharff <aramzs@hacktext.com>",
+	"auto": {
+		"plugins": [
+			"all-contributors",
+			"first-time-contributor",
+			"released",
+			"git-tag"
+		],
+		"onlyPublishWithReleaseLabel": true
+	}
 }
